third_party/WebKit/LayoutTests/fast/forms/select-popup/popup-menu-touch-operations-expected.txt

PASS picker._selectElement.value is "1"
PASS menuElement.value is "bar"
PASS picker._selectElement.value is "0"
PASS picker._selectElement.value is "2"
PASS internals.pagePopupWindow is null
PASS menuElement.value is "baz"
==> Test disabled option
PASS picker._selectElement.value is "2"
PASS menuElement.value is "baz"
PASS picker._selectElement.value is "2"
PASS picker._selectElement.value is "2"
PASS internals.pagePopupWindow is not null
PASS menuElement.value is "baz"
==> Test scrollable popup
PASS picker._selectElement.value is "1"
PASS menuElement2.value is "2"
PASS picker._selectElement.value is "2"
PASS picker._selectElement.value is "3"
PASS picker._selectElement.value is "0"
PASS internals.pagePopupWindow is not null
PASS menuElement2.value is "2"
==> Test popup closes on outside GestureTapDawn
PASS internals.pagePopupWindow is null
==> Test popup doesn't reopen immediately after closing
PASS internals.pagePopupWindow is null
PASS internals.pagePopupWindow is null
PASS internals.pagePopupWindow is not null
PASS internals.pagePopupWindow is null
PASS internals.pagePopupWindow is null
PASS successfullyParsed is true

TEST COMPLETE
 
